requirements.txt
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-chroma
langchain-groq
chromadb
pypdf
python-dotenv
streamlit
numpy<2.0
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-chroma
langchain-groq
chromadb
pypdf
python-dotenv
streamlit
numpy<2.0
sentence-transformers